factor out app setup code into separate folder modules
this change focuses on reorganizing where setup code is located by
moving some of it to separate folders. By putting setup code in separate
folders, it hides from index.js how these aspects are implemented and
what packages they use. this follows the dependency inversion principle
because the concrete implementation is hidden behind a generic
interface, such as `controllers.startup()` and `models.startup()` (look
for these methods in this code change).

* change name of jwtAuthMiddleware to validateJwtMiddleware and update all
controllers that use it.
* move openapi-enforcer-middleware code into the controllers module and
define two new exports, `middleware` and `startup()`
* factor out final error handler into its own file at the top-level of
the project
* move `sequelize.authenticate()` and sequelize error handling
middleware into the models module as `startup()` and
`errorHandlerMiddleware`
* move swagger documentation serving into its own module,
swaggerDocsRouter
