+++ controllers/index.js
@@ -2,9 +2,18 @@ const users = require("./users");
 const messages = require("./messages");
 const likes = require("./likes");
 const auth = require("./auth");
+
+const EnforcerMiddleware = require("openapi-enforcer-middleware");
+const enforcer = EnforcerMiddleware("./specification.yaml");
+
 module.exports = {
   ...auth,
   ...likes,
   ...messages,
-  ...users
+  ...users,
+  middleware: enforcer.middleware(),
+  startup: async () => {
+    await enforcer.promise;
+    await enforcer.controllers("./controllers");
+  }
 };
